tempo: remove beat quantization, use Gaussian BP
Changing to a quantized (to 0.25 BPM) tempo in beat histogram, and
then changing back to a quantized (integer) samples in beat phase
doesn't make sense.

To deal with slight variations in sample-based tempi, the
cumulative Beat Phase now adds a gaussian curve instead of a
single point for each tick.